[fix] fix code share info
before, if the flight_number_element has <img> tag, it means that the flight is code share flight,
but, acutually, the <img> tags are added not only code share flight but also NFC flight.
so, to find code share info or not, change to check <a> tag in the flight_number_element.