[feat]Output format of 会社別RJTT2RJCC定時到着率.py round-trip
This commit modifies the `main` function in `会社別RJTT2RJCC定時到着率.py` to allow the user to specify the departure and destination airports as command line arguments. If no arguments are provided, the default route between HND and CTS is used. Additionally, the script now calculates the on-time arrival rate for both the specified route and the reverse route. The changes also include adding a print statement at the end of the file.